HACK: for some reason some roms keep hotplugging cpu1. Stop that.
This also affects oc apps, which now will not be able to hotplug cpu1
to change the frequency on cpu1. So the only way to get a frequency
change on both cpus 'on-the-fly': Starting a benchmark (cpu1 comes up) and then set
the freq.

The preferred method should be flashing the kernel with the right khz
values from recovery.
I hope that the problem is solved soon so this can get reverted eventually.
(it appears to be fixed in newer cm9 nightlies)

Signed-off-by: showp1984 <showp1984@gmail.com>