[PATCH] Expose virtual PID from child namespace in /proc/PID/status.
Date: Wed, 27 Feb 2013 17:34:33 -0500
Subject: [PATCH] Expose virtual PID from child namespace in /proc/PID/status.

The virtual PID is used by Ubuntu Touch application manager code ( which
runs in the Android namespace ), to communicate with code that runs inside the
Ubuntu container.

This patch is not from upstream.  It was proposed on the lxc-users
mailing list:

http://www.mail-archive.com/lxc-users@lists.sourceforge.net/msg01840.html

Change-Id: If63652002c227827fb524e03a1d0210fc2b7d4b1
Original-Author: Greg Kurz <gkurz@fr.ibm.com>
Signed-off-by: Greg Kurz <gkurz@fr.ibm.com>
Signed-off-by: Tony Espy <espy@canonical.com>
Signed-off-by: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com>